订单删除

+++ app/src/main/java/com/yikangcheng/admin/yikang/bean/ClassifyCommodityListBean.java
@@ -149,8 +149,8 @@ public class ClassifyCommodityListBean {
         private int isavaliable;
         private String addtime;
         private int isPay;
-        private int sourceprice;
-        private int currentprice;
+        private double sourceprice;
+        private double currentprice;
         private String title;
         private String logo;
         private int losetype;
@@ -265,19 +265,19 @@ public class ClassifyCommodityListBean {
             this.isPay = isPay;
         }
 
-        public int getSourceprice() {
+        public double getSourceprice() {
             return sourceprice;
         }
 
-        public void setSourceprice(int sourceprice) {
+        public void setSourceprice(double sourceprice) {
             this.sourceprice = sourceprice;
         }
 
-        public int getCurrentprice() {
+        public double getCurrentprice() {
             return currentprice;
         }
 
-        public void setCurrentprice(int currentprice) {
+        public void setCurrentprice(double currentprice) {
             this.currentprice = currentprice;
         }
